requirements.txt
flask==3.0.3
flask-cors==4.0.1
transformers==4.44.2
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
